pyproject.toml
[project]
name = "llm-evolution"
version = "0.1.0"
description = "LLM + evolutionary algorithms to optimize programs in multiple languages and domains"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "loky>=3.5.6",
    "openai>=2.14.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "README.md",
    "pyproject.toml",
    "src/llm_evolution",
]

[tool.hatch.build.targets.wheel]
packages = ["src/llm_evolution"]

[tool.basedpyright]
exclude = [
    "tests",
    "**/__pycache__"
]

[tool.pytest.ini_options]
addopts = "-v --capture=no"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
